Fixed the casing of @Json parameter
Update and Insert were recently broken because the InstanceToParameters was changed to add the parameter `@JSON` but the SQL statement was still referring to `@Json`. This works fine when the Server Collation of the SQL Server is CI, but breaks when the Server Collation is CS (regardless of the Database Collation). See the related issue for more details.

Fixes https://github.com/OctopusDeploy/Issues/issues/3734