Share a table alias generator for a whole transaction.
This means that multiple queries in a single transaction may influence each other (in terms of the aliases used in the queries). As long as the queries that we actually run in a transaction are predictable, then this should be fine.
It fixes an issue where combining multiple queries into one query (eg using joins) could result in alias collisions.